# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LaraBook API",
        "description": "LaraBook API",
        "contact": {
            "email": "hubertlipinskipl@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api/",
            "description": "Localhost server"
        }
    ],
    "paths": {
...
        "/book/search": {
            "get": {
                "tags": [
                    "book"
                ],
                "summary": "Search book list",
                "description": "Search book list",
                "operationId": "search",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Pagination page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Pagination page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "$ref": "#/components/parameters/book_search_id"
                    },
                    {
                        "$ref": "#/components/parameters/book_search_title"
                    },
                    {
                        "$ref": "#/components/parameters/book_search_description"
                    },
                    {
                        "$ref": "#/components/parameters/book_search_short_description"
                    },
                    {
                        "$ref": "#/components/parameters/book_search_user_id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/BookCollection"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "Password Based": []
                    }
                ]
            }
        },
        "/oauth/token": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Login",
                "description": "Login",
                "operationId": "Login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "grant_type",
                                    "client_id",
                                    "client_secret"
                                ],
                                "properties": {
                                    "grant_type": {
                                        "type": "string",
                                        "example": "password"
                                    },
                                    "client_id": {
                                        "type": "string",
                                        "example": "id"
                                    },
                                    "client_secret": {
                                        "type": "string",
                                        "example": "secret"
                                    },
                                    "username": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        },
                                        "expires_in": {
                                            "type": "integer",
                                            "example": "31536000"
                                        },
                                        "access_token": {
                                            "type": "string",
                                            "example": "access_token"
                                        },
                                        "refresh_token": {
                                            "type": "string",
                                            "example": "refresh_token"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "BookCollection": {
                "title": "BookCollection",
                "description": "Book collection",
                "properties": {
                    "books": {
                        "title": "books",
                        "description": "Array of books",
                        "type": "array",
                        "format": "array",
                        "items": {
                            "$ref": "#/components/schemas/BookResource"
                        }
                    },
                    "pagination": {
                        "$ref": "#/components/schemas/PaginationResource"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "BookCollection"
                }
            },
            "BookResource": {
                "title": "BookResource",
                "description": "Book resource",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "Book id",
                        "format": "integer",
                        "example": 1
                    },
                    "title": {
                        "title": "title",
                        "description": "Book title",
                        "format": "string",
                        "example": "Title"
                    },
                    "description": {
                        "title": "description",
                        "description": "Book description",
                        "format": "string",
                        "example": "Description"
                    },
                    "short_description": {
                        "title": "short_description",
                        "description": "Book short description",
                        "format": "string",
                        "example": "Short description"
                    },
                    "created_at": {
                        "title": "created_at",
                        "description": "User created at",
                        "format": "string",
                        "example": "2021-09-06T20:58:20.000000Z"
                    },
                    "updated_at": {
                        "title": "created_at",
                        "description": "User updated at",
                        "format": "string",
                        "example": "2021-09-06T20:58:20.000000Z"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "BookResource"
                }
            },
            "PaginationResource": {
                "title": "PaginationResource",
                "description": "Pagination resource",
                "properties": {
                    "total": {
                        "title": "total",
                        "description": "Total number of items in the data store",
                        "format": "int64",
                        "example": "0"
                    },
                    "count": {
                        "title": "count",
                        "description": "Number of items for the current page",
                        "format": "int64",
                        "example": 0
                    },
                    "per_page": {
                        "title": "per_page",
                        "description": "Number of items to be shown per page",
                        "format": "int64",
                        "example": 10
                    },
                    "current_page": {
                        "title": "current_page",
                        "description": "Current page number",
                        "format": "int64",
                        "example": 1
                    },
                    "total_pages": {
                        "title": "total_pages",
                        "description": "Last available page number",
                        "format": "int64",
                        "example": 0
                    },
                    "links": {
                        "properties": {
                            "previous_page": {
                                "type": "string",
                                "example": "null"
                            },
                            "next_page": {
                                "type": "string",
                                "example": "null"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "PaginationResource"
                }
            },
...
        },
        "parameters": {
            "book_search_id": {
                "name": "id",
                "in": "query",
                "description": "Filter books by id",
                "required": false,
                "schema": {
                    "type": "integer"
                }
            },
            "book_search_title": {
                "name": "title",
                "in": "query",
                "description": "Filter books by title",
                "required": false,
                "schema": {
                    "type": "string"
                }
            },
            "book_search_description": {
                "name": "description",
                "in": "query",
                "description": "Filter books by description",
                "required": false,
                "schema": {
                    "type": "string"
                }
            },
            "book_search_short_description": {
                "name": "short_description",
                "in": "query",
                "description": "Filter books by short description",
                "required": false,
                "schema": {
                    "type": "string"
                }
            },
            "book_search_user_id": {
                "name": "user_id",
                "in": "query",
                "description": "Filter books by owner id",
                "required": false,
                "schema": {
                    "type": "integer"
                }
            }
        },
...
        "securitySchemes": {
            "Password Based": {
                "type": "oauth2",
                "description": "Authenticate to API",
                "in": "header",
                "flows": {
                    "password": {
                        "tokenUrl": "/oauth/token",
                        "scopes": {}
                    }
                },
                "scheme": "token"
            }
        }
    },
    "security": [
        []
    ]
}